add week2 work
