+++ src/App.test.ts
@@ -1,9 +1,10 @@
 import { render, screen } from '@testing-library/react';
 
+import React from 'react';
 import App from './App';
 
 test('App', () => {
-  render(<App />);
+  render(React.createElement(App, null));
 
   screen.getByText(/Hello, wholeman!/);
 });
