+++ src/main.ts
@@ -1,3 +1,4 @@
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 
 import App from './App';
@@ -9,7 +10,8 @@ function main() {
   }
 
   const root = ReactDOM.createRoot(container);
-  root.render(<App />);
+  const app = React.createElement(App, null, null);
+  root.render(app);
 }
 
 main();
